backend/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
numpy
pandas
scikit-learn
pymongo
motor
python-dateutil
requests
aiohttp
aiohttp_socks 